Remove MiniPortile from extension build
The extension build currently downloads and installs sources for
libiconv, openssl, and freetds regardless of architecture. This is at
odds with the correct procedure of installing these sources via the OS
package manager for the non-binary gem build.

This commit begins to solve this problem in a few ways:

1. It moves the responsibiliy of managing the MiniPortile builds to a
   ports rake task
2. It removes all ports specific logic from the extension build save the
   ability to search the ports directories if they exist (to facilitate
   the binary gem)
3. It adds cross compilation targets for the ports rake tasks to assist
   in the creation of the binary gems.
4. It enhances the rake clobber list to include the ports directories.
5. It removes custom logic for ports directory names and moves back
   towards relying on the MiniPortile defaults as much as possible.

Relates to issue #342